+++ deps/smoldot.ts
@@ -1,5 +1,7 @@
-export * from "https://deno.land/x/smoldot2@light-js-deno-v1.0.6/index-deno.js"
+export * from "https://deno.land/x/smoldot2@light-js-deno-v1.0.11/index-deno.js"
 export type {
+  AddChainOptions,
+  Chain,
   Client,
   ClientOptions,
-} from "https://deno.land/x/smoldot2@light-js-deno-v1.0.6/public-types.d.ts"
+} from "https://deno.land/x/smoldot2@light-js-deno-v1.0.11/public-types.d.ts"
