+++ rpc/known/mod.ts
@@ -4,6 +4,7 @@ import { BeefyCalls, BeefySubscriptions } from "./beefy.ts"
 import { ChainCalls, ChainSubscriptions } from "./chain.ts"
 import { ChildStateCalls } from "./childstate.ts"
 import { ContractsCalls } from "./contracts.ts"
+import { ExperimentalCalls, ExperimentalSubscriptions } from "./experimental_spec.ts"
 import { FrameSystemCalls } from "./framesystem.ts"
 import { GrandpaCalls, GrandpaSubscriptions } from "./grandpa.ts"
 import { MmrCalls } from "./mmr.ts"
@@ -30,6 +31,7 @@ export type Calls =
   & SyncCalls
   & SystemCalls
   & TransactionPaymentCalls
+  & ExperimentalCalls
 
 export type Subscriptions =
   & AuthorSubscriptions
@@ -37,6 +39,7 @@ export type Subscriptions =
   & ChainSubscriptions
   & GrandpaSubscriptions
   & StateSubscriptions
+  & ExperimentalSubscriptions
 
 export * from "./author.ts"
 export * from "./babe.ts"
@@ -44,12 +47,12 @@ export * from "./beefy.ts"
 export * from "./chain.ts"
 export * from "./childstate.ts"
 export * from "./contracts.ts"
+export * from "./experimental_spec.ts"
 export * from "./framesystem.ts"
 export * from "./grandpa.ts"
 export * from "./mmr.ts"
 export * from "./offchain.ts"
 export * from "./payment.ts"
-export * as smoldot from "./smoldot.ts"
 export * from "./state.ts"
 export * from "./statemigration.ts"
 export * from "./sync.ts"
